added tutorial prompt and tutorial button
