+++ src/Sandbox/Sandbox.jsx
@@ -8,13 +8,13 @@ import ToolWindow from "./UI/ToolWindow/ToolWindow";
 import InfoWindow from "./UI/InfoWindow/InfoWindow";
 import Hierarchy from "./UI/Hierarchy/Hierarchy";
 import ColorWindow from "./UI/ColorWindow/ColorWindow";
-import NameTag from "./UI/NameTag/NameTag";
+import NameArea from "./UI/NameArea/NameArea";
 import LoadScreen from "../UI/LoadScreen/LoadScreen";
 import ModeBar from "./UI/ModeBar/ModeBar";
 import WarningPop from "../UI/WarningPop/WarningPop";
 import ProjectDim from "./UI/ProjectDim/ProjectDim";
 import MouseHover from "./UI/MouseHover/MouseHover";
-import Tutorial from "./UI/Tutorial/Tutorial";
+import TutorialPrompt from "./UI/NameArea/Tutorial/TutorialPrompt";
 
 import KeyTracker from "./DevTools/KeyTracker";
 import ProjectReader from "./DevTools/ProjectReader";
@@ -132,7 +132,7 @@ const Sandbox = () => {
 
   return (
     <>
-      <Tutorial />
+      <TutorialPrompt />
       <MouseHover ref={mouseHoverRef} />
       <div className="webpage">
         <WarningPop />
@@ -146,7 +146,7 @@ const Sandbox = () => {
         />
         <Banner />
         <AuthTester reroute={"/"} />
-        <NameTag />
+        <NameArea />
         <div className="sandbox">
           <Exporter ref={exporterRef} screenshotRef={screenshotRef} />
           <KeyTracker />
