+++ src/Sandbox/UI/NameArea/Tutorial/tutorial.css
@@ -1,0 +1,76 @@
+.tutorial-container{
+    display:flex;
+    justify-content:center;
+    align-items:center;
+    z-index:1000;
+    height:120vh;
+    width:100vw;
+    position:absolute;
+}
+.tutorial-blackout{
+    height:100%;
+    width:100%;
+    background-color:rgba(0, 0, 0,0.5);
+    position:absolute;
+}
+.tutorial-text-container{
+    display:flex;
+    flex-direction:column;
+    align-items:center;
+    justify-content:center;
+    z-index:1;
+    background-color:rgb(153, 185, 219);
+    border-radius:25px;
+    height:50vh;
+    width:40vw;
+    border-color:rgb(125, 100, 198);
+    border-style:solid;
+}
+
+.tutorial-text{
+    color:rgb(90, 105, 160);
+    font-family:"Nunito",sans-serif;
+    font-weight:700;
+    font-size:5vh;
+    text-align:center;
+    margin-bottom:5vh;
+    margin-top:5vh;
+}
+.tutorial-options-container{
+    display:flex;
+    flex-direction:row;
+    justify-content:space-between;
+    align-items:center;
+    margin-bottom:10vh;
+    width:75%;
+}
+.tutorial-prev {
+    display:flex;
+    justify-content:center;
+    align-items:center;
+    color:white;
+    height:10vh;
+    width:30vh;
+    background-color:rgb(88, 94, 201);
+    border-radius:100px;
+    font-size:3vh;
+    font-family:"Nunito",sans-serif;
+    font-weight:600;
+}
+
+.tutorial-next {
+    display:flex;
+    justify-content:center;
+    align-items:center;
+    color:rgb(88, 94, 201);
+    height:10vh;
+    width:30vh;
+    background-color:white;
+    border-radius:100px;
+    font-size:3vh;
+    font-family:"Nunito",sans-serif;
+    font-weight:700;
+}
+.tutorial-button{
+    height: 4vh;
+}
